requirements.txt
streamlit
google-generativeai
PyMuPDF
python-dotenv
